+++ ipmi-oem/ipmi-oem-inventec.c
@@ -1542,7 +1542,7 @@ ipmi_oem_inventec_set_board_id (ipmi_oem_state_data_t *state_data)
   uint8_t bytes_rq[IPMI_OEM_MAX_BYTES];
   uint8_t bytes_rs[IPMI_OEM_MAX_BYTES];
   uint8_t boardid;
-  unsigned int tmp;
+  unsigned long tmp;
   char *endptr;
   int rs_len;
   int rv = -1;
@@ -1567,7 +1567,7 @@ ipmi_oem_inventec_set_board_id (ipmi_oem_state_data_t *state_data)
                        state_data->prog_data->args->oem_options[0]);
       goto cleanup;
     }
-  boardid = tmp;
+  boardid = (uint8_t)tmp;
 
   /* Inventec 5441/Dell Xanadu II OEM
    * Inventec 5442/Dell Xanadu III OEM
@@ -1686,7 +1686,7 @@ ipmi_oem_inventec_set_fcb_version (ipmi_oem_state_data_t *state_data)
   uint8_t bytes_rs[IPMI_OEM_MAX_BYTES];
   uint8_t majorversion;
   uint8_t minorversion;
-  unsigned int tmp;
+  unsigned long tmp;
   char *endptr;
   int rs_len;
   int rv = -1;
@@ -1711,7 +1711,7 @@ ipmi_oem_inventec_set_fcb_version (ipmi_oem_state_data_t *state_data)
                        state_data->prog_data->args->oem_options[0]);
       goto cleanup;
     }
-  majorversion = tmp;
+  majorversion = (uint8_t)tmp;
 
   errno = 0;
 
@@ -1730,7 +1730,7 @@ ipmi_oem_inventec_set_fcb_version (ipmi_oem_state_data_t *state_data)
                        state_data->prog_data->args->oem_options[1]);
       goto cleanup;
     }
-  minorversion = tmp;
+  minorversion = (uint8_t)tmp;
 
   /* Inventec 5441/Dell Xanadu II OEM
    * Inventec 5442/Dell Xanadu III OEM
@@ -2087,7 +2087,7 @@ ipmi_oem_inventec_set_sol_inactivity_timeout (ipmi_oem_state_data_t *state_data)
   if (strcasecmp (state_data->prog_data->args->oem_options[0], "none"))
     {
       char *endptr = NULL;
-      unsigned int temp;
+      unsigned long temp;
 
       errno = 0;
 
@@ -2105,7 +2105,7 @@ ipmi_oem_inventec_set_sol_inactivity_timeout (ipmi_oem_state_data_t *state_data)
           goto cleanup;
         }
 
-      sol_inactivity_timeout = temp;
+      sol_inactivity_timeout = (uint16_t)temp;
     }
   else
     sol_inactivity_timeout = 0;
